+++ forui/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.21.0 (Next)
+
+### `FPopoverMenu`
+* Add `FPopoverMenuStyle.minWidth`.
+
+
+
+
 ## 0.20.0
 
 This update builds upon the styling overhaul by introducing desktop and touch variants of each theme and making widget